requirements.txt
pypdf==6.4.2
fpdf2==2.8.1
